# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_0xsj_nexus-go_internal_email_application_dto.ActivateTemplateResponse:
    properties:
      template:
        $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.TemplateDTO'
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.ArchiveTemplateResponse:
    properties:
      template:
        $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.TemplateDTO'
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.CreateTemplateRequest:
    properties:
      body_html:
        type: string
      body_text:
        type: string
      description:
        type: string
      locale:
        type: string
      name:
        type: string
      slug:
        type: string
      subject:
        type: string
      tenant_id:
        type: string
      variables:
        items:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.VariableRequest'
        type: array
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.CreateTemplateResponse:
    properties:
      template:
        $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.TemplateDTO'
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.DeleteTemplateResponse:
    properties:
      id:
        type: string
      success:
        type: boolean
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.GetTemplateResponse:
    properties:
      template:
        $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.TemplateDTO'
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.ListTemplatesResponse:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      templates:
        items:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.TemplateDTO'
        type: array
      total:
        type: integer
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.PreviewTemplateBySlugRequest:
    properties:
      data:
        additionalProperties: true
        type: object
      locale:
        type: string
      slug:
        type: string
      tenant_id:
        type: string
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.PreviewTemplateResponse:
    properties:
      body_html:
        type: string
      body_text:
        type: string
      subject:
        type: string
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.TemplateDTO:
    properties:
      body_html:
        type: string
      body_text:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      locale:
        type: string
      name:
        type: string
      slug:
        type: string
      status:
        type: string
      subject:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      variables:
        items:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.VariableDTO'
        type: array
      version:
        type: integer
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.UpdateTemplateRequest:
    properties:
      body_html:
        type: string
      body_text:
        type: string
      description:
        type: string
      name:
        type: string
      subject:
        type: string
      variables:
        items:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.VariableRequest'
        type: array
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.UpdateTemplateResponse:
    properties:
      template:
        $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.TemplateDTO'
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.VariableDTO:
    properties:
      default:
        type: string
      description:
        type: string
      name:
        type: string
      required:
        type: boolean
      type:
        type: string
    type: object
  github_com_0xsj_nexus-go_internal_email_application_dto.VariableRequest:
    properties:
      default:
        type: string
      description:
        type: string
      name:
        type: string
      required:
        type: boolean
      type:
        type: string
    type: object
...
  internal_email_interface_http_v1.ErrorResponse:
    description: Error response returned by the API
    properties:
      code:
        example: TEMPLATE_NOT_FOUND
        type: string
      message:
        example: template not found
        type: string
      meta:
        additionalProperties: {}
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Multi-tenant API built with hexagonal architecture. Provides identity
    management, tenant management, email templates, and more.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Hexagonal Go API
  version: "1.0"
paths:
...
  /api/v1/email/templates:
    get:
      description: Retrieves a paginated list of email templates with optional filters
      parameters:
      - description: Filter by tenant ID
        format: uuid
        in: query
        name: tenant_id
        type: string
      - description: Filter by status
        enum:
        - draft
        - active
        - archived
        - deleted
        in: query
        name: status
        type: string
      - description: Filter by locale
        example: en
        in: query
        name: locale
        type: string
      - description: Filter by slug containing text
        in: query
        name: slug_contains
        type: string
      - description: Filter by name containing text
        in: query
        name: name_contains
        type: string
      - default: true
        description: Include system templates
        in: query
        name: include_system
        type: boolean
      - default: 0
        description: Pagination offset
        in: query
        minimum: 0
        name: offset
        type: integer
      - default: 20
        description: Pagination limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Sort field
        enum:
        - created_at
        - updated_at
        - name
        - slug
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of templates
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.ListTemplatesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List email templates
      tags:
      - email
    post:
      consumes:
      - application/json
      description: Creates a new email template with subject, body, and variable definitions
      parameters:
      - description: Create template request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Template created successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.CreateTemplateResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "409":
          description: Template with slug already exists
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create email template
      tags:
      - email
  /api/v1/email/templates/{id}:
    delete:
      description: Soft-deletes an email template
      parameters:
      - description: Template ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Template deleted successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.DeleteTemplateResponse'
        "400":
          description: Invalid template ID
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete email template
      tags:
      - email
    get:
      description: Retrieves an email template by its unique identifier
      parameters:
      - description: Template ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Template found
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.GetTemplateResponse'
        "400":
          description: Invalid template ID
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get email template by ID
      tags:
      - email
    put:
      consumes:
      - application/json
      description: Updates an existing email template
      parameters:
      - description: Template ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Update template request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.UpdateTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Template updated successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.UpdateTemplateResponse'
        "400":
          description: Invalid request body or template ID
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update email template
      tags:
      - email
  /api/v1/email/templates/{id}/activate:
    post:
      description: Activates a draft or archived template, making it available for
        use
      parameters:
      - description: Template ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Template activated successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.ActivateTemplateResponse'
        "400":
          description: Invalid template ID
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "422":
          description: Invalid status transition
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Activate email template
      tags:
      - email
  /api/v1/email/templates/{id}/archive:
    post:
      description: Archives an active template, removing it from active use
      parameters:
      - description: Template ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Template archived successfully
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.ArchiveTemplateResponse'
        "400":
          description: Invalid template ID
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "422":
          description: Invalid status transition
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Archive email template
      tags:
      - email
  /api/v1/email/templates/{id}/preview:
    post:
      consumes:
      - application/json
      description: Renders an email template with provided data for preview
      parameters:
      - description: Template ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Template variables data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Rendered template preview
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.PreviewTemplateResponse'
        "400":
          description: Invalid request body or template ID
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Preview email template
      tags:
      - email
  /api/v1/email/templates/by-slug:
    get:
      description: Retrieves an email template by its slug and locale
      parameters:
      - description: Template slug
        example: welcome-email
        in: query
        name: slug
        required: true
        type: string
      - default: en
        description: Locale code
        example: en
        in: query
        name: locale
        type: string
      - description: Tenant ID
        format: uuid
        in: query
        name: tenant_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Template found
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.GetTemplateResponse'
        "400":
          description: Slug is required
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get email template by slug
      tags:
      - email
  /api/v1/email/templates/preview-by-slug:
    post:
      consumes:
      - application/json
      description: Renders an email template by slug with provided data for preview
      parameters:
      - description: Preview request with slug and data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.PreviewTemplateBySlugRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rendered template preview
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_email_application_dto.PreviewTemplateResponse'
        "400":
          description: Invalid request body or missing slug
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/internal_email_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Preview email template by slug
      tags:
      - email
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...